+++ app/Http/Controllers/Api/V1/UserController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Models\Task;
+use App\Models\Task\Task;
 use App\Models\User\User;
 use App\Repositories\UserRepository;
 use Illuminate\Http\Request;
