+++ app/Models/Task/Task.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Task;
 
 use Eloquent;
